Pretty headings on customisation files and tidied up.
I noticed someone who was using figlet to produce attractive headers
in their text files comments, and I liked the idea. I had to be
careful about how this was done though as the time-stamp function only
works if the placeholder is in the first 8 lines of the file.

On this basic I have moved to a single line of comment characters,
then a time-stamp placeholder, then an empty line and then the file
name in figlet 'bubble writing'.

After the bubble writing file name I have added comments as appropriate.